requirements.txt
# To ensure app dependencies are ported from your virtual environment/host machine into your container, run 'pip freeze > requirements.txt' in the terminal to overwrite this file
django>=4.0, <4.1.0
gunicorn>=20.1.0, <20.2.0

djangorestframework>=3.13.1,<3.14.0